+++ src/models/Settings.php
@@ -36,11 +36,6 @@ class Settings extends Model
     /**
      * @var string
      */
-    public string $soundcloudKey = '';
-    
-    /**
-     * @var string
-     */
     public string $facebookKey = '';
 
     /**
@@ -232,7 +227,7 @@ class Settings extends Model
         return [
             'parser' => [
                 'class' => EnvAttributeParserBehavior::class,
-                'attributes' => ['embedlyKey', 'iframelyKey', 'googleKey', 'soundcloudKey', 'facebookKey', 'referer'],
+                'attributes' => ['embedlyKey', 'iframelyKey', 'googleKey', 'facebookKey', 'referer'],
             ],
         ];
     }
@@ -243,7 +238,7 @@ class Settings extends Model
     protected function defineRules(): array
     {
         return [
-            [['embedlyKey', 'iframelyKey', 'googleKey', 'soundcloudKey', 'facebookKey', 'referer'], StringValidator::class],
+            [['embedlyKey', 'iframelyKey', 'googleKey', 'facebookKey', 'referer'], StringValidator::class],
             ['parameters', 'each', 'rule' => [ParameterValidator::class]],
             [['whitelist', 'extraWhitelist'], 'each', 'rule' => [StringValidator::class]],
             [['maxAssetNameLength', 'maxFileNameLength'], 'integer', 'min' => 10],